+++ photutils/background/__init__.py
@@ -1,0 +1,8 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This subpackage contains modules and packages for background and
+background rms estimation.
+"""
+
+from .core import *
+from .background_2d import *